bug fix in libscpi sample code: change netconn_write() argument from NETCONN_NOCOPY to NETCONN_COPY to fix corrupted output when using chained queries and single queries with multiple return values. NETCONN_COPY can only be used in cases like *IDN?, which is probably why it hasn't been discovered in test cases.
See netconn_write() documentation: "When passed the flag NETCONN_COPY the data is copied into internal buffers which are allocated for the data. This allows the data to be modified directly after the call, but is inefficient both in terms of execution time and memory usage. If the flag is not set then the data is not copied but rather referenced, and the NETCONN_NOCOPY manifest is provided for backwards compatibility. *The data must not be modified after the call*, since the data can be put on the retransmission queue for the connection, and stay there for an indeterminate amount of time."